+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/hdinsight/spark/ui/SparkBatchJobConfigurable.java
@@ -32,7 +32,6 @@ import com.microsoft.azure.hdinsight.spark.common.SparkBatchJobConfigurableModel
 import com.microsoft.azure.hdinsight.spark.common.SparkSubmitModel;
 import com.microsoft.azuretools.azurecommons.helpers.NotNull;
 import com.microsoft.azuretools.telemetry.TelemetryConstants;
-import com.microsoft.intellij.telemetry.IdeaTelemetryUtils;
 
 import javax.swing.*;
 import java.awt.*;
@@ -53,13 +52,12 @@ public class SparkBatchJobConfigurable implements SettableControl<SparkBatchJobC
     private final Project myProject;
 
     @NotNull
-    protected String getServiceName() {
+    public String getServiceName() {
         return TelemetryConstants.HDINSIGHT;
     }
 
     public SparkBatchJobConfigurable(@NotNull final Project project) {
         this.myProject = project;
-        IdeaTelemetryUtils.INSTANCE.addTelemetryListener(getComponent(), getServiceName());
     }
 
     @NotNull
