+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/hdinsight/spark/ui/SparkLocalRunParamsPanel.java
@@ -34,8 +34,6 @@ import com.intellij.ui.DocumentAdapter;
 import com.intellij.ui.MacroAwareTextBrowseFolderListener;
 import com.intellij.util.ui.UIUtil;
 import com.microsoft.azure.hdinsight.spark.common.SparkLocalRunConfigurableModel;
-import com.microsoft.azuretools.telemetry.TelemetryConstants;
-import com.microsoft.intellij.telemetry.IdeaTelemetryUtils;
 import org.apache.commons.lang3.SystemUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
